+++ src/main/java/com/pluton/yelody/serviceImpl/ChartServiceImpl.java
@@ -4,16 +4,18 @@ import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
+import org.hibernate.exception.ConstraintViolationException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import com.pluton.yelody.exceptions.ConstraintViolationHandler;
 import com.pluton.yelody.exceptions.EntityNotFoundException;
+import com.pluton.yelody.exceptions.UniqueEntityException;
 import com.pluton.yelody.models.Chart;
 import com.pluton.yelody.repositories.ChartRepository;
 import com.pluton.yelody.services.ChartService;
-import com.pluton.yelody.utilities.ImageUtil;
 
 @Service
 public class ChartServiceImpl implements ChartService {
@@ -27,23 +29,20 @@ public class ChartServiceImpl implements ChartService {
 	public ResponseEntity<Object> postChart(Chart chart){
 		try {
 			return new ResponseEntity<Object>(chartRepository.save(chart), HttpStatus.CREATED);
-		}catch(Exception ex) {
-			return new ResponseEntity<Object>(HttpStatus.FOUND);
-		}
+		}catch(Exception e) {
+			 if (e.getCause() instanceof ConstraintViolationException) {
+		            ConstraintViolationException constraintViolationException = (ConstraintViolationException) e.getCause();
+		            String duplicateValue = constraintViolationException.getSQLException().getMessage();
+		            throw new UniqueEntityException(duplicateValue);
+		        } else {
+		            throw new UniqueEntityException(e.getMessage());
+		        }
+			}
 	}
 
 	@Override
 	public Optional<Chart> getChartById(UUID id) {
-		chartGet = null;
-		try {
-			chartGet = Optional.ofNullable(chartRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("CHART ID: " + id + " NOT FOUND")));
-			if(chartGet!=null)
-				return chartGet;
-
-			return null;
-		}catch(Exception ex) {
-			return null;
-		}
+			return Optional.ofNullable(chartRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("CHART ID: " + id + " NOT FOUND")));
 	}
 
 	@Override
@@ -61,14 +60,12 @@ public class ChartServiceImpl implements ChartService {
 	}
 
 	@Override
-	public ResponseEntity<?> deleteChart(Chart chart) {
-		try {
-			chartRepository.delete(chart);
-			ImageUtil.deleteFile(chart.getImage());
-			return ResponseEntity.status(HttpStatus.OK).body("CHART " + chart.getName() + " HAS BEEN DELETED SUCCESSFULLY");
-		}catch(Exception ex){
-			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("");
-		}
+	public HttpStatus deleteChart(Chart chart) {
+		if(chart.getSongs()!=null && !chart.getSongs().isEmpty())
+			throw new ConstraintViolationHandler("Constraint violation: This CHART is associated with SONG(s) and cannot be deleted.");
+
+		chartRepository.delete(chart);
+		return HttpStatus.OK;
 	}
